# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProFile API",
    "description": "ProFile Resume & Portfolio Management API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    }
  ],
...
  "paths": {
    "/api/v1/resumes": {
      "get": {
        "operationId": "resumes_getAllUserResumes",
        "tags": [
          "resumes"
        ],
        "summary": "Get all resumes for current user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get all resumes for current user"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "post": {
        "operationId": "resumes_createResumeForUser",
        "tags": [
          "resumes"
        ],
        "summary": "Create a new resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Create a new resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/resumes/{resumeId}/collaborators": {
      "get": {
        "operationId": "collaboration_getCollaborators",
        "tags": [
          "collaboration"
        ],
        "summary": "Get collaborators for a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get collaborators for a resume"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "post": {
        "operationId": "collaboration_invite",
        "tags": [
          "collaboration"
        ],
        "summary": "Invite user to collaborate on resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCollaboratorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invite user to collaborate on resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaboratorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/resumes/shared-with-me": {
      "get": {
        "operationId": "collaboration_getSharedWithMe",
        "tags": [
          "collaboration"
        ],
        "summary": "Get resumes shared with current user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get resumes shared with current user"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/resumes/{resumeId}/collaborators/{userId}": {
      "patch": {
        "operationId": "collaboration_updateRole",
        "tags": [
          "collaboration"
        ],
        "summary": "Update collaborator role",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update collaborator role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaboratorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "delete": {
        "operationId": "collaboration_remove",
        "tags": [
          "collaboration"
        ],
        "summary": "Remove collaborator from resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove collaborator from resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/users/me/data-export": {
      "get": {
        "operationId": "gdpr_exportUserData",
        "tags": [
          "gdpr"
        ],
        "summary": "Export all user data (GDPR Right to Access)",
        "description": "Downloads all user data in machine-readable JSON format as required by GDPR Article 20.",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Export all user data (GDPR Right to Access)"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/users/me/account": {
      "delete": {
        "operationId": "gdpr_deleteEndpoint",
        "tags": [
          "gdpr"
        ],
        "summary": "Delete user account (GDPR Right to be Forgotten)",
        "description": "Permanently deletes the user account and all associated data as required by GDPR Article 17. This action is irreversible.",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Delete user account (GDPR Right to be Forgotten)"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/auth/change-email": {
      "post": {
        "operationId": "auth_changeEmail",
        "tags": [
          "auth"
        ],
        "summary": "Change email address (authenticated)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Change email address (authenticated)"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/auth/delete-account": {
      "post": {
        "operationId": "auth_deleteAccount",
        "tags": [
          "auth"
        ],
        "summary": "Delete own account (authenticated)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Delete own account (authenticated)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "DeleteResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "ResumeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "title": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "example": "false"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "language": {
            "type": "string"
          },
          "targetRole": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "isPublic",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "InviteCollaboratorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "user-uuid-v4",
            "description": "User ID to invite as collaborator"
          },
          "role": {
            "type": "string",
            "example": "EDITOR",
            "description": "Role to assign to the collaborator"
          }
        },
        "required": [
          "userId",
          "role"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "VIEWER",
            "description": "New role for the collaborator"
          }
        },
        "required": [
          "role"
        ]
      },
      "CollaboratorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "userId": {
            "type": "string",
            "example": "clxxx..."
          },
          "resumeId": {
            "type": "string",
            "example": "clxxx..."
          },
          "role": {
            "type": "string",
            "example": "EDITOR"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "resumeId",
          "role",
          "createdAt"
        ]
      },
...
    }
  }
}